% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log2.R
\name{log2}
\alias{log2}
\title{log2}
\usage{
log2(x)
}
\arguments{
\item{x}{input value}
}
\value{
log(x) for non-negative x, and trigger errors with a negative input
}
\description{
log2
}
\examples{
x=10
log2(x)
}
