% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrt2.R
\name{sqrt2}
\alias{sqrt2}
\title{sqrt2}
\usage{
sqrt2(x)
}
\arguments{
\item{x}{input value}
}
\value{
square root of x for non-negative x, and trigger errors with a negative input
}
\description{
sqrt2
}
\examples{
x=10
sqrt2(x)
}
